Move sticky logic to globalTemplate to add padding to the container
