+++ components/layouts/globalTemplate.module.css
@@ -1,0 +1,3 @@
+.stickyPageWrapper {
+  @apply pt-24;
+}
